Organize fluent folder
